init: add project settings
