+++ nodemon.json
@@ -1,0 +1,6 @@
+{
+  "watch": ["src", "./index.ts"],
+  "ext": ".ts,.js",
+  "ignore": [],
+  "exec": "npx ts-node ./index.ts"
+}
